requirements.txt
libclang
pyyaml